backend/requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.3
Flask-Login==0.2.9
Flask-Security==1.7.1
Flask-HTTPAuth==2.2.0
twill==0.9
simplejson==3.3.2
Flask-Testing==0.4
